% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_slash_to_path.R
\name{add_slash_to_path}
\alias{add_slash_to_path}
\title{add \code{/} to the path if needed}
\usage{
add_slash_to_path(data_source)
}
\arguments{
\item{data_source}{Character. path}
}
\value{
Path with \code{/} at the end
}
\description{
add \code{/} to the path if needed
}
