% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange_plotlist.R
\name{arrange_plotlist}
\alias{arrange_plotlist}
\title{Arrange list of ggplots into single figure}
\usage{
arrange_plotlist(
  plotlist,
  ncol = NULL,
  nrow = NULL,
  labels = NULL,
  widths = 1,
  heights = 1,
  legend = NULL,
  common_legend = FALSE
)
}
\arguments{
\item{plotlist}{List of plots to display.}

\item{ncol}{(optional) number of columns in the plot grid.}

\item{nrow}{(optional) number of rows in the plot grid.}

\item{labels}{(optional) list of labels to be added to the plots. You can
also set labels="AUTO" to auto-generate upper-case labels or labels="auto"
to auto-generate lower-case labels.}

\item{widths}{(optional) numerical vector of relative columns widths. For
example, in a two-column grid, widths = c(2, 1) would make the first column
twice as wide as the second column.}

\item{heights}{same as \code{widths} but for column heights.}

\item{legend}{character specifying legend position. Allowed values are one of
c("top", "bottom", "left", "right", "none"). To remove the legend use
legend = "none".}

\item{common_legend}{logical value. Default is FALSE. If TRUE, a common
unique legend will be created for arranged plots.}
}
\description{
A warpper around \code{ggpubr::ggarrange()}
}
\seealso{
\code{\link[ggpubr:ggarrange]{ggpubr::ggarrange()}}
}
