% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_class.R
\name{check_class}
\alias{check_class}
\title{Assert the class of object}
\usage{
check_class(data_source, sel_class)
}
\arguments{
\item{data_source}{The name of the object in quotes}

\item{sel_class}{The name of the class in quotes}
}
\value{
\code{TRUE} if class match or error message
}
\description{
The function will evaluate the object of \code{data_source} name
and test if ANY of the classes in equal to \code{sel_class}
}
