% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_col_names.r
\name{check_col_names}
\alias{check_col_names}
\title{Check the presence of variables in data.frame}
\usage{
check_col_names(data_source, var_list)
}
\arguments{
\item{data_source}{Name of the data.frame to check}

\item{var_list}{Vector with names}
}
\description{
Function will test the presence of ALL names in \code{var_list}
within the \code{data_source} data.frame and return error message
}
