% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_color.R
\name{check_color}
\alias{check_color}
\title{Check if the selected object is a color}
\usage{
check_color(data_source)
}
\arguments{
\item{data_source}{Character. Name of the vector to check}
}
\description{
The vector can contain both name of a color or a hex value
}
\seealso{
\code{\link[=is_color]{is_color()}}
}
