% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_if_integer.R
\name{check_if_integer}
\alias{check_if_integer}
\title{Check if object is and integer}
\usage{
check_if_integer(data_source)
}
\arguments{
\item{data_source}{The name of the object in quotes}
}
\value{
\code{TRUE} if class match or error message
}
\description{
Check if object is and integer
}
