% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_if_loaded.R
\name{check_if_loaded}
\alias{check_if_loaded}
\title{Confirm that the selected file is present in selected environment}
\usage{
check_if_loaded(file_name, env = rlang::current_env, verbose = TRUE)
}
\arguments{
\item{file_name}{Name of the selected file}

\item{env}{The environment to test the presence of the file in}

\item{verbose}{Logical. Should there be output message?}
}
\description{
Confirm that the selected file is present in selected environment
}
