% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_files.R
\name{compare_files}
\alias{compare_files}
\title{Compare two datasets}
\usage{
compare_files(file_a, file_b, file_a_envir = NULL, file_b_envir = NULL)
}
\arguments{
\item{file_a}{The object A to compare}

\item{file_b}{The object B to compare}
}
\value{
Logical. Are those file the same?
}
\description{
Compare the two files and return if same
}
