% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confirm_choice.R
\name{confirm_choice}
\alias{confirm_choice}
\title{Open menu to confirm TRUE/FALSE}
\usage{
confirm_choice(msg = "", default_value = TRUE)
}
\arguments{
\item{msg}{Message to be shown with the menu}

\item{default_value}{Default value of answer}
}
\value{
return logical
}
\description{
Open menu to confirm TRUE/FALSE
}
