% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_in_time.R
\name{do_in_time}
\alias{do_in_time}
\title{Evaluate an expression and interrupts it if it takes too long}
\usage{
do_in_time(expr, envir = parent.frame(), timeout)
}
\arguments{
\item{expr}{The R expression to be evaluated}

\item{envir}{The environment in which the expression should be evaluated}

\item{timeout}{A numeric specifying the maximum number of seconds the
expression is allowed to run before being interrupted by the \code{timeout}}
}
\description{
Evaluate an expression and interrupts it if it takes too long
}
\details{
Function heavily instipred by \code{\link[R.utils:withTimeout]{R.utils::withTimeout()}}
}
