% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_var_from_list.R
\name{extract_var_from_list}
\alias{extract_var_from_list}
\title{Helper function to extract data safely}
\usage{
extract_var_from_list(var, dataset)
}
\arguments{
\item{var}{Name of the variable}

\item{dataset}{A list}
}
\description{
Helper function to extract data safely
}
