% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten_list_by_one.R
\name{flatten_list_by_one}
\alias{flatten_list_by_one}
\title{Flatten the list by one level}
\usage{
flatten_list_by_one(data_source)
}
\arguments{
\item{data_source}{A list}
}
\description{
Return list without the first h-level. The names of the h1 and
h2 are merged
}
