% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clean_name.R
\name{get_clean_name}
\alias{get_clean_name}
\title{get a name of clean name of file}
\usage{
get_clean_name(data_source)
}
\arguments{
\item{data_source}{Character. Name of a file}
}
\value{
Name of a file without date and SHA code
}
\description{
get a name of clean name of file
}
