% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_date_from_name.R
\name{get_date_from_name}
\alias{get_date_from_name}
\title{Get date from file name}
\usage{
get_date_from_name(data_source)
}
\arguments{
\item{data_source}{Character with name of a file}
}
\value{
Character with date
}
\description{
Get date from file name
}
