% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_format_from_name.R
\name{get_format_from_name}
\alias{get_format_from_name}
\title{Get format of  file name}
\usage{
get_format_from_name(data_source)
}
\arguments{
\item{data_source}{string with file name}
}
\value{
String with format
}
\description{
Get format of  file name
}
