% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_latest_file.R
\name{get_latest_file}
\alias{get_latest_file}
\title{Detect the newest version of the selected file and load it}
\usage{
get_latest_file(file_name, dir = here::here(), verbose = TRUE)
}
\arguments{
\item{file_name}{Name of the object to save in quotes}

\item{dir}{Directory path}

\item{verbose}{Logical. Should there be output message?}
}
\value{
Object of latest version of the \code{file_name}
}
\description{
Look into the folder \code{dir} and find the version of the file with
the most recent date and load it
}
