% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_latest_name_file.R
\name{get_latest_name_file}
\alias{get_latest_name_file}
\title{Detect the newest version of the selected file}
\usage{
get_latest_name_file(file_name, dir, silent = FALSE)
}
\arguments{
\item{file_name}{Name of the object to save in quotes}

\item{dir}{Directory path}

\item{silent}{Logical. Should error message be ouput}
}
\value{
Object name of the most recent file
}
\description{
look into the \code{dir} folder and find the version of the file
with the most recent name
}
