% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_latest_name_folder.R
\name{get_latest_name_folder}
\alias{get_latest_name_folder}
\title{Detect the newest version of the selected folder}
\usage{
get_latest_name_folder(folder_name, dir)
}
\arguments{
\item{folder_name}{Name of the object to save in quotes}

\item{dir}{Directory path}
}
\value{
Object name of the most recent folder
}
\description{
look into the \code{dir} folder and find the version of the folder
with the most recent name
}
