% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sha_from_name.R
\name{get_sha_from_name}
\alias{get_sha_from_name}
\title{Get SHA from file name}
\usage{
get_sha_from_name(data_source)
}
\arguments{
\item{data_source}{Character with name of a file}
}
\value{
Character with SHA code
}
\description{
Get SHA from file name
}
