% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_format.R
\name{guess_format}
\alias{guess_format}
\title{Guess format of  file name}
\usage{
guess_format(data_source)
}
\arguments{
\item{data_source}{string with file name}
}
\value{
String with format
}
\description{
Guess format of  file name
}
