% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_color.R
\name{is_color}
\alias{is_color}
\title{Test if a character is a color}
\usage{
is_color(data_source)
}
\arguments{
\item{data_source}{Character with name of a color. Can be aslo a hex value}
}
\value{
Logical
}
\description{
Test if a character is a color
}
