% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_comment.R
\name{output_comment}
\alias{output_comment}
\title{Output message in console}
\usage{
output_comment(msg = "")
}
\arguments{
\item{msg}{String with the message that should be printed}
}
\description{
Output message in console
}
