% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_heading.R
\name{output_heading}
\alias{output_heading}
\title{Output boxed message in console}
\usage{
output_heading(msg = "", size = c("h1", "h2", "h3"))
}
\arguments{
\item{msg}{String with the message that should be printed}

\item{size}{Select the hierarchy of the heading (1-3)}
}
\description{
Output boxed message in console
}
