% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_warning.R
\name{output_warning}
\alias{output_warning}
\title{Output warning in console}
\usage{
output_warning(msg = "")
}
\arguments{
\item{msg}{String with the message that should be printed}
}
\description{
Output warning in console
}
