% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste_as_vector.R
\name{paste_as_vector}
\alias{paste_as_vector}
\title{Paste any list of values as a vector}
\usage{
paste_as_vector(var_list, sep = "'")
}
\arguments{
\item{var_list}{List of values}

\item{sep}{Separator which should be used to wrap each object in
the \code{var_list}}
}
\description{
Paste any list of values as a vector
}
