% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_null_with_na.R
\name{replace_null_with_na}
\alias{replace_null_with_na}
\title{Replace NULL with N}
\usage{
replace_null_with_na(data_source)
}
\arguments{
\item{data_source}{Any R object}
}
\description{
Helper function. If \code{data_source} is \code{NULL} then replace with \code{NA}
}
