% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUtilpol-package.R
\docType{package}
\name{RUtilpol-package}
\alias{RUtilpol}
\alias{RUtilpol-package}
\title{RUtilpol: R-Utilpol: Utility Functions For Package Development}
\description{
R-Utilpol is an R package proving utility functions for package development.
}
\author{
\strong{Maintainer}: Ondrej Mottl \email{ondrej.mottl@gmail.com}

}
\keyword{internal}
