% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_latest_file.R
\name{save_latest_file}
\alias{save_latest_file}
\title{Save the newest version of data}
\usage{
save_latest_file(
  file_name,
  dir,
  current_date = Sys.Date(),
  prefered_format = c("qs", "csv"),
  preset = c("high", "balanced", "fast", "uncompressed", "archive")
)
}
\arguments{
\item{file_name}{Character. Name of the object to save in quotes}

\item{dir}{Character. Directory path}

\item{current_date}{Character. Current date}

\item{prefered_format}{Character. format to save as: "qs" or "csv"}

\item{preset}{Charatcer. One of "fast", "balanced", "high" (default), "archive",
"uncompressed". See `qs::qsave()`` for details}
}
\description{
Look into the folder and find the version of the file with
the most recent name. Compare the last saved file and selected file and
save file if something changed since recent.
Note that if file has changes and it has the same date as \code{current_date},
the file will be overwritten (old file deleted)
}
\seealso{
\code{\link[qs:qsave]{qs::qsave()}}
}
