% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip_files.R
\name{zip_files}
\alias{zip_files}
\title{Compress Files into 'zip'}
\usage{
zip_files(
  zipfile,
  files,
  recurse = TRUE,
  include_directories,
  root = ".",
  mode = c("cherry-pick", "mirror")
)
}
\arguments{
\item{zipfile}{The zip file to create. If the file exists, \code{zip}
overwrites it, but \code{zip_append} appends to it.}

\item{files}{List of file to add to the archive. See details below
about absolute and relative path names.}

\item{recurse}{Whether to add the contents of directories recursively.}

\item{include_directories}{Whether to explicitly include directories
in the archive. Including directories might confuse MS Office when
reading docx files, so set this to \code{FALSE} for creating them.}

\item{root}{Change to this working directory before creating the
archive.}

\item{mode}{Selects how files and directories are stored in
the archive. It can be \code{"mirror"} or \code{"cherry-pick"}.
See "Relative Paths" in \code{\link[zip:zip]{zip::zip()}}.}
}
\description{
A wrappper function to a \code{zip::zip()}
}
\seealso{
\code{\link[zip:zip]{zip::zip()}}
}
