% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_var_from_list.R
\name{extract_var_from_list}
\alias{extract_var_from_list}
\title{Helper function to extract data safely}
\usage{
extract_var_from_list(data_source_var, data_source_list)
}
\arguments{
\item{data_source_var}{Name of the data_source_variable}

\item{data_source_list}{A list}
}
\description{
Helper function to extract data safely
}
