% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_latest_file_name.R
\name{get_latest_file_name}
\alias{get_latest_file_name}
\title{Detect the newest version of the selected file}
\usage{
get_latest_file_name(
  file_name,
  dir = here::here(),
  folder = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{file_name}{Name of the object to save in quotes}

\item{dir}{Directory path}

\item{folder}{Logical. Is the selected file a folder?}

\item{verbose}{Logical. Should there be output error message?}
}
\value{
Object name of the most recent file
}
\description{
look into the \code{dir} folder and find the version of the file
with the most recent name
}
